# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Competitor Insight Hub API",
    "description": "API для мониторинга новостей из мира ИИ-индустрии",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Current User Profile",
        "description": "Get current user profile",
        "operationId": "get_current_user_profile_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update Current User",
        "description": "Update current user profile",
        "operationId": "update_current_user_api_v1_users_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/users/preferences": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Preferences",
        "description": "Get user preferences",
        "operationId": "get_user_preferences_api_v1_users_preferences_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User Preferences",
        "description": "Update user preferences",
        "operationId": "update_user_preferences_api_v1_users_preferences_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_frequency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Notification Frequency"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_user_preferences_api_v1_users_preferences_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    },
    "/api/v1/users/companies/{company_id}/subscribe": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Subscribe To Company",
        "description": "Subscribe to a company",
        "operationId": "subscribe_to_company_api_v1_users_companies__company_id__subscribe_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    },
    "/api/v1/users/companies/{company_id}/unsubscribe": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Unsubscribe From Company",
        "description": "Unsubscribe from a company",
        "operationId": "unsubscribe_from_company_api_v1_users_companies__company_id__unsubscribe_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    },
    "/api/v1/users/preferences/digest": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Digest Settings",
        "description": "Get user digest settings",
        "operationId": "get_digest_settings_api_v1_users_preferences_digest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update Digest Settings",
        "description": "Update user digest settings",
        "operationId": "update_digest_settings_api_v1_users_preferences_digest_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DigestSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_update_user_preferences_api_v1_users_preferences_put": {
        "properties": {
          "subscribed_companies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subscribed Companies"
          },
          "interested_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interested Categories"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "type": "object",
        "title": "Body_update_user_preferences_api_v1_users_preferences_put"
      },
...
      "DigestSettingsUpdate": {
        "properties": {
          "digest_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest Enabled"
          },
          "digest_frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest Frequency"
          },
          "digest_custom_schedule": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest Custom Schedule"
          },
          "digest_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest Format"
          },
          "digest_include_summaries": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest Include Summaries"
          },
          "telegram_chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Chat Id"
          },
          "telegram_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Enabled"
          },
          "telegram_digest_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Digest Mode"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "week_start_day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Start Day"
          }
        },
        "type": "object",
        "title": "DigestSettingsUpdate",
        "description": "Model for updating digest settings"
      },
...
      "UserUpdateSchema": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name",
            "description": "User's full name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "Whether user account is active"
          }
        },
        "type": "object",
        "title": "UserUpdateSchema",
        "description": "Schema for user updates"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}